+++ mpmorph/workflow/converge.py
@@ -1,5 +1,3 @@
-from fireworks import explicit_serialize, FireTaskBase, FWAction, Workflow
-from pymatgen import Structure
 from mpmorph.fireworks import powerups
 from mpmorph.fireworks.core import MDFW, OptimizeFW
 from mpmorph.util import recursive_update
@@ -262,18 +260,3 @@ def get_converge_new(structure, temperature, converge_scheme='EOS', priority=Non
                                                     db_file=run_args["run_specs"]["db_file"])
     return fw_list
 
-
-@explicit_serialize
-class DiffusionTask(FireTaskBase):
-    required_params = ['temperatures', 'max_steps', 'target_steps', 'trajectory_to_db']
-    optional_params = []
-
-    def run_task(self, fw_spec):
-        s = Structure.from_file('CONTCAR.gz')
-        fws = []
-        for t in self['temperatures']:
-            fws.extend(get_converge_new(s, t, max_steps=self['max_steps'],
-                                           target_steps=self['target_steps'],
-                                           trajectory_to_db=self['trajectory_to_db']))
-        wf = Workflow(fws)
-        return FWAction(detours=wf)
